requirements.txt
fastapi
uvicorn
firebase-admin
python-jose
requests
